startup_funding/requirements.txt
python-dotenv==1.0.1
browser-use==0.1.36
libsql-client==0.3.1
pydantic==2.10.6
openai==1.61.1
selenium==4.28.1